Conflicts:
	practice_project/DAL/src/main/resources/META-INF/persistence.xml
	practice_project/jsf_module/src/main/webapp/WEB-INF/faces-config.xml
	practice_project/jsf_module/src/main/webapp/index.xhtml
	practice_project/pom.xml

+++ practice_project/DAL/src/main/java/org/DAL/CommentService.java
@@ -43,7 +43,7 @@ public class CommentService extends CommentDAO {
 
 	@Override
 	@TransactionAttribute(TransactionAttributeType.REQUIRED)
-	public void removeComment(int сommentId) {
+	public void removeComment(long сommentId) {
 		super.removeComment(сommentId);
 	}
 }
